Separate DOM code from utility classes
